refactor(frontend); remove ignore action type
