+++ frontend/app/src/components/history/IgnoredInAccountingIcon.vue
