+++ frontend/app/src/types/history/ignored.ts
@@ -1,8 +1,3 @@
-export enum IgnoreActionType {
-  HISTORY_EVENTS = 'history_event',
-}
-
 export interface IgnorePayload {
-  actionType: IgnoreActionType.HISTORY_EVENTS;
   data: string[];
 }
